moving problematic test to the Full stack harness
